feat(arcs): finalize arc link labels
