+++ packages/arcs/src/arc_link_labels/ArcLinkLabelsLayer.tsx
@@ -1,4 +1,5 @@
 import React from 'react'
+import { usePropertyAccessor } from '@nivo/core'
 import { DatumWithArcAndColor } from '../types'
 import { useArcLinkLabelsTransition } from './useArcLinkLabelsTransition'
 import { ArcLinkLabelsProps } from './props'
@@ -26,7 +27,7 @@ interface ArcLinkLabelsLayerProps<Datum extends DatumWithArcAndColor> {
 export const ArcLinkLabelsLayer = <Datum extends DatumWithArcAndColor>({
     center,
     data,
-    label,
+    label: labelAccessor,
     skipAngle,
     offset,
     diagonalLength,
@@ -37,9 +38,15 @@ export const ArcLinkLabelsLayer = <Datum extends DatumWithArcAndColor>({
     linkColor,
     component = ArcLinkLabel,
 }: ArcLinkLabelsLayerProps<Datum>) => {
-    const { transition, interpolateLink } = useArcLinkLabelsTransition<Datum>({
+    const getLabel = usePropertyAccessor<Datum, string>(labelAccessor)
+
+    const {
+        transition,
+        interpolateLink,
+        interpolateTextAnchor,
+        interpolateTextPosition,
+    } = useArcLinkLabelsTransition<Datum>({
         data,
-        label,
         skipAngle,
         offset,
         diagonalLength,
@@ -57,6 +64,7 @@ export const ArcLinkLabelsLayer = <Datum extends DatumWithArcAndColor>({
                 return React.createElement(Label, {
                     key: datum.id,
                     datum,
+                    label: getLabel(datum),
                     style: {
                         ...transitionProps,
                         thickness: strokeWidth,
@@ -68,6 +76,22 @@ export const ArcLinkLabelsLayer = <Datum extends DatumWithArcAndColor>({
                             transitionProps.offset,
                             transitionProps.diagonalLength,
                             transitionProps.straightLength
+                        ),
+                        textAnchor: interpolateTextAnchor(
+                            transitionProps.startAngle,
+                            transitionProps.endAngle,
+                            transitionProps.innerRadius,
+                            transitionProps.outerRadius
+                        ),
+                        textPosition: interpolateTextPosition(
+                            transitionProps.startAngle,
+                            transitionProps.endAngle,
+                            transitionProps.innerRadius,
+                            transitionProps.outerRadius,
+                            transitionProps.offset,
+                            transitionProps.diagonalLength,
+                            transitionProps.straightLength,
+                            transitionProps.textOffset
                         ),
                     },
                 })
