+++ packages/arcs/src/arc_link_labels/index.ts
@@ -1,4 +1,8 @@
 export * from './ArcLinkLabelsLayer'
 export * from './canvas'
+export * from './compute'
 export * from './props'
+export * from './types'
+export * from './useArcLinkLabels'
 export * from './useArcLinkLabelsTransition'
+export * from './useArcLinks'
