+++ packages/arcs/src/index.ts
@@ -7,7 +7,6 @@ export * from './boundingBox'
 export * from './centers'
 export * from './interactivity'
 export * from './interpolateArc'
-export * from './links'
 export * from './types'
 export * from './useAnimatedArc'
 export * from './useArcGenerator'
